+++ src/main/java/com/xueersi/parentsmeeting/modules/livevideoOldIJK/business/courseware/CoursewarePreload.java
@@ -110,6 +110,8 @@ public class CoursewarePreload {
                 //buglys上面有报Attempt to get length of null array,加上try,catch
                 try {
                     File[] files = file.listFiles();
+                    if (files == null)
+                        return;
                     for (File itemFile : files) {
                         //文件夹是日期格式并且不是今天才删除
                         if (isCoursewareDir(itemFile.getName()) && !itemFile.getName().equals(today)) {
@@ -146,6 +148,7 @@ public class CoursewarePreload {
             return;
         }
         File[] files = file.listFiles();
+        if (files == null) return;
         for (File itemFile : files) {
             if (!itemFile.isDirectory()) {
                 itemFile.delete();
